NeighborlyAPI/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Josh Haines", email = "Josh@JoshHaines.com"},
]
dependencies = [
    "pymongo==3.10.1",
    "azure-functions==1.2.1",
    "icecream~=2.1",
]
requires-python = ">=3.9"
dynamic = ["classifiers"]
license = {text = "MIT"}

[project.urls]
homepage = ""

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"
verify_ssl = true

[tool.pdm.dev-dependencies]
lint = [
    "black"
]
